+++ test/test-osvvm.tcl
@@ -6,14 +6,22 @@
 if {[info exists env(OSVVM_DIR)]} {
     set OsvvmDir $env(OSVVM_DIR)
 } else {
-    # known good snapshot
-    set OsvvmLibraries_tag "2023.09a"
-    set OsvvmDir [file join $::env(HOME) .cache nvc "OsvvmLibraries-${OsvvmLibraries_tag}"]
+    set scriptdir [ file dirname [ file normalize [ info script ] ] ]
+    set fid [open "$scriptdir/../contrib/install-osvvm.sh" r]
+    while {[gets $fid line] != -1} {
+        if {[regexp -all -- {git_wrapper \S+ (\S+)} $line -> tag]} {
+            break
+        }
+    };
+    close $fid;
+
+    set OsvvmDir [file join $::env(HOME) .cache nvc "OsvvmLibraries-$tag"]
 }
 
 source $OsvvmDir/Scripts/StartNVC.tcl
 
 #SetVHDLVersion 2019
+SetExtendedElaborateOptions {--jit}
 
 set ::osvvm::AnalyzeErrorStopCount 1
 set ::osvvm::SimulateErrorStopCount 1